+++ springcore/src/main/java/com/affiasco/springcore/SpringcoreApplication.java
@@ -6,8 +6,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class SpringcoreApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(SpringcoreApplication.class, args);
-	}
-
+    public static void main(String[] args) {
+        SpringApplication.run(SpringcoreApplication.class, args);
+    }
 }
