+++ springcore/src/main/java/com/affiasco/springcore/common/Coach.java
@@ -1,4 +1,4 @@
-package com.affiasco.springcore;
+package com.affiasco.springcore.common;
 
 public interface Coach {
     String getDailyWorkout();
