+++ springcore/src/main/java/com/affiasco/springcore/rest/DemoController.java
@@ -1,5 +1,6 @@
-package com.affiasco.springcore;
+package com.affiasco.springcore.rest;
 
+import com.affiasco.springcore.common.Coach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
